requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
Pillow>=10.0.0
flask>=3.0.0
flask-cors>=4.0.0
tqdm>=4.65.0
opencv-python>=4.8.0
